Crate aw_fel

Allwinner FEL library.

Modules

soc
SoC module.

uboot

Structs

Fel
Allwinner FEL devices context.

FelHandle
FEL device handle.

UsbHandle
USB device handle.

Enums

FelError
FEL errors.

Constants

AW_FEL_1_EXEC
Allwinner FEL execution request.

AW_FEL_1_READ
Allwinner FEL read request.

AW_FEL_1_WRITE
Allwinner FEL write request.

AW_FEL_VERSION
Allwinner FEL version request.

AW_PRODUCT_ID
Allwinner FEL device USB product ID.

AW_USB_MAX_BULK_SEND
AW_USB_MAX_BULK_SEND and the timeout constant USB_TIMEOUT are related. Both need to be selected in a way that transferring the maximum chunk size with (SoC-specific) slow transfer speed won't time out.

AW_USB_READ
Allwinner USB read request.

AW_USB_WRITE
Allwinner USB write request.

AW_VENDOR_ID
Allwinner FEL device USB vendor ID.

DRAM_BASE
DRAM base address.

DRAM_SIZE
DRAM size, in bytes.

LCODE_ARM_RMR_WORDS
Word count of the rmr_request scratch code.

LCODE_ARM_RW_SIZE
Code size in bytes.

LCODE_ARM_RW_WORDS
Word count of the [read/write]_words() scratch code.

LCODE_MAX_RW_WORDS
Data words for read/write requests.

LCODE_MAX_TOTAL
Maximum total words in buffer.

SPL_LEN_LIMIT
Maximum size of SPL, at the same time this is the start offset of the main U-Boot image within u-boot-sunxi-with-spl.bin.

USB_TIMEOUT
USB timeout (in seconds).

_DERIVE_failure_Fail_FOR_FelError

_DERIVE_failure_core_fmt_Display_FOR_FelError

Functions

u32_as_u8
Converts a reference to a u32 slice to a reference to a u8 slice.

u32_as_u8_mut
Converts a mutable reference to a u32 slice to a mutable reference to a u8 slice.