Struct aw_fel::soc::Info

pub struct Info {
    soc_id: u32,
    name: &'static str,
    spl_addr: u32,
    scratch_addr: u32,
    thunk_addr: u32,
    thunk_size: u32,
    needs_l2en: bool,
    mmu_tt_addr: Option<u32>,
    sid_addr: Option<u32>,
    rvbar_reg: Option<u32>,
    swap_buffers: &'static [SRAMSwapBuffers],
}

SoC information structure.

Each SoC variant may have its own list of memory buffers to be exchanged and the information about the placement of the thunk code, which handles the transition of execution from the BROM FEL code to the U-Boot SPL and back.

Note: the entries in the swap_buffers tables need to be sorted by buf1 addresses. And the buf1 addresses are the BROM data buffers, while buf2 addresses are the intended backup locations.

Also for performance reasons, we optionally want to have MMU enabled with optimal section attributes configured (the code from the BROM should use I-cache, writing data to the DRAM area should use write combining). The reason is that the BROM FEL protocol implementation moves data using the CPU somewhere on the performance critical path when transferring data over USB. The older SoC variants (A10/A13/A20/A31/A23) already have MMU enabled and we only need to adjust section attributes. The BROM in newer SoC variants (A33/A83T/H3) doesn't enable MMU any more, so we need to find some 16K of spare space in SRAM to place the translation table there and specify it as the mmu_tt_addr field in the soc_sram_info structure. The mmu_tt_addr address must be 16K aligned.

Fields
soc_id: u32

ID of the SoC.

name: &'static str

Human-readable SoC name string.

spl_addr: u32

SPL load address.

scratch_addr: u32

A safe place to upload & run code.

thunk_addr: u32

Address of the thunk code.

thunk_size: u32

Maximal size of the thunk code.

needs_l2en: bool

Set the L2EN bit.

mmu_tt_addr: Option<u32>

MMU translation table address.

sid_addr: Option<u32>

Base address for SID_KEY[0-3] registers.

rvbar_reg: Option<u32>

MMIO address of RVBARADDR0_L register.

swap_buffers: &'static [SRAMSwapBuffers]

SRAM buffers.

Methods

impl Info

pub fn from_id(soc_id: u32) -> Option<Self>

Gets the SoC information structure for the given ID, if supported.

pub fn from_version(version: &Version) -> Option<Self>

Gets the SoC information structure from the givern SoCVersion, if supported.

pub fn get_soc_id(&self) -> u32

Gets the SoC ID.

pub fn get_name(&self) -> &str

Gets the name of the SoC.

pub fn get_spl_addr(&self) -> u32

Gets the SPL load address for the SoC.

pub fn get_scratch_addr(&self) -> u32

Gets the scratch address.

pub fn get_thunk_addr(&self) -> u32

Gets the address of the thunk code.

pub fn get_thunk_size(&self) -> u32

Gets the size of the thunk code.

pub fn needs_l2en(&self) -> bool

pub fn get_mmu_tt_addr(&self) -> Option<u32>

Gets the MMU translation table address.

pub fn get_sid_addr(&self) -> Option<u32>

Gets the SID register address.

pub fn get_rvbar_reg(&self) -> Option<u32>

Get the RVBAR register address, if supported.

pub fn get_swap_buffers(&self) -> &[SRAMSwapBuffers]

Gets the SoC swap buffers.