Struct aw_fel::soc::SRAMSwapBuffers

pub struct SRAMSwapBuffers {
    buf1: u32,
    buf2: u32,
    size: u32,
}

SRAM buffers.

The SRAMSwapBuffers structure is used to describe information about pairwise memory regions in SRAM, the content of which needs to be exchanged before calling the U-Boot SPL code and then exchanged again before returning control back to the FEL code from the BROM.

Fields
buf1: u32

BROM buffer.

buf2: u32

Backup storage location.

size: u32

Buffer size.

Methods

impl SRAMSwapBuffers

pub fn get_buf1(&self) -> u32

Gets the BROM buffer.

pub fn get_buf2(&self) -> u32

Gets the backup storage location.

pub fn get_size(&self) -> u32

Gets the buffer size.