Struct aw_fel::FelHandle

pub struct FelHandle<'h> {
    usb_handle: UsbHandle<'h>,
    soc_version: Version,
    soc_info: Info,
}

FEL device handle.

Fields
usb_handle: UsbHandle<'h> soc_version: Version soc_info: Info
Methods

impl<'h> FelHandle<'h>

pub fn write_uboot_image(&self, uboot: &[u8]) -> Result<(u32, u32), Error>

This function tests a given buffer for a valid U-Boot image. Upon success, the image data gets transferred to the default memory address stored within the image header; and the function returns the U-Boot entry point (offset) and size values, in that order.

pub fn write_and_execute_spl(&self, spl: &[u8]) -> Result<(), Error>

Writes U-Boot SPL into memory and executes it.

impl<'h> FelHandle<'h>

pub fn get_soc_info(&self) -> &Info

Gets the SoC information from the FEL device.

Note: This is a no-op. The SoC information is acquired during initialization.

pub fn get_version_info(&self) -> &Version

Gets the SoC version information from the FEL device.

Note: This is a no-op. The SoC information is acquired during initialization.

fn enable_l2_cache(&self) -> Result<(), Error>

Enables the L2 cache.

fn get_stack_info(&self) -> Result<(u32, u32), Error>

Gets stack information.

The result will be a tuple. The first element will be the SP_irq, while the second element will be the nuser mode SP.

fn default_mmu_translation_table() -> [u32; 4096]

Generates a default MMU translation table.

fn backup_and_disable_mmu(&self) -> Result<Option<[u32; 4096]>, Error>

Backup MMU translation table and disable it.

fn restore_and_enable_mmu(&self, tt: [u32; 4096]) -> Result<(), Error>

fn get_ttbr0(&self) -> Result<u32, Error>

Gets the TTBR0 register.

fn set_ttbr0(&self, val: u32) -> Result<(), Error>

Sets the TTBR0 register to the given value.

fn get_ttbcr(&self) -> Result<u32, Error>

Gets the TTBCR register.

fn set_ttbcr(&self, val: u32) -> Result<(), Error>

Sets the TTBCR register to the given value.

fn get_dacr(&self) -> Result<u32, Error>

Gets the DACR register.

fn set_dacr(&self, val: u32) -> Result<(), Error>

Sets the DACR register to the given value.

fn get_sctlr(&self) -> Result<u32, Error>

Gets the SCTLR register.

fn set_sctlr(&self, val: u32) -> Result<(), Error>

Sets the SCTLR register to the given value.

fn read_arm_cp_reg(
&self,
coproc: u32,
opc1: u32,
cr_n: u32,
cr_m: u32,
opc2: u32
) -> Result<u32, Error>

Reads the given ARM register.

fn write_arm_cp_reg(
&self,
coproc: u32,
opc1: u32,
cr_n: u32,
cr_m: u32,
opc2: u32,
val: u32
) -> Result<(), Error>

Writes the given value to the given ARM register.

pub fn read_sid(&self) -> Result<Option<[u32; 4]>, Error>

Reads the SID from the SoC if it has one.

pub fn read_words(&self, offset: u32, words: &mut [u32]) -> Result<(), Error>

Read words chunk from the FEL device memory.

pub fn write_words(&self, offset: u32, words: &[u32]) -> Result<(), Error>

Write words chunk to the FEL device memory.

fn read_words_chunk(&self, offset: u32, words: &mut [u32]) -> Result<(), Error>

Read words chunk from the FEL device memory.

fn write_words_chunk(&self, offset: u32, words: &[u32]) -> Result<(), Error>

Write words chunk to the FEL device memory.

pub fn rmr_request(&self, entry_point: u32, aarch64: bool) -> Result<(), Error>

Perform FEL RMR request.

This function stores a given entry point to the RVBAR address for CPU0, and then writes the Reset Management Register to request a warm boot. It is useful with some AArch64 transitions, e.g. when passing control to ARM Trusted Firmware (ATF) during the boot process of Pine64.

The code was inspired by U-boot