Struct aw_fel::UsbHandle

pub struct UsbHandle<'h> {
    device_handle: DeviceHandle<'h>,
    endpoint_in: u8,
    endpoint_out: u8,
    iface_detached: bool,
}

USB device handle.

Fields
device_handle: DeviceHandle<'h> endpoint_in: u8 endpoint_out: u8 iface_detached: bool
Methods

impl<'h> UsbHandle<'h>

fn from_device(device: &Device<'h>) -> Result<UsbHandle<'h>, Error>

Creates a USB handle from the given device.

fn claim(&mut self, device: &Device<'h>) -> Result<(), Error>

Claim the given device.

fn get_endpoints(&mut self, device: &Device<'h>) -> Result<(), Error>

Updates the endpoints of the USB handle.

pub fn fel_fill(
&self,
offset: u32,
num_bytes: u32,
byte: u8
) -> Result<(), Error>

Fill memory at the given offset.

It will fill num_size bytes with the given byte.

pub fn fel_read(&self, offset: u32, buf: &mut [u8]) -> Result<(), Error>

Reads from the FEL device memory at the given offset.

It will fill all the given buffer, but it will panic if the buffer overflows total memory address space.

pub fn fel_write(&self, offset: u32, buf: &[u8]) -> Result<(), Error>

Writes to the FEL device memory at the given offset.

It will panic if the buffer overflows total memory address space.

pub fn fel_execute(&self, offset: u32) -> Result<(), Error>

Makes the FEL device execute the code at the given address.

fn get_fel_version(&self) -> Result<Version, Error>

Gets the SoC version information.

fn read_fel_status(&self) -> Result<(), Error>

Reads the FEL status.

fn send_fel_request(
&self,
req_type: u32,
addr: u32,
len: u32
) -> Result<(), Error>

Sends a FEL request.

fn usb_read(&self, data: &mut [u8]) -> Result<(), Error>

Reads the given data from the USB.

fn usb_write(&self, data: &[u8]) -> Result<(), Error>

Writes the given data to the USB.

fn send_usb_request(&self, req_type: u16, len: u32) -> Result<(), Error>

Sends the request type USB request.

fn read_usb_response(&self) -> Result<(), Error>

Reads and discards the USB response. It will check if the response is correct.

fn usb_bulk_send(
device_handle: &DeviceHandle,
endpoint: u8,
data: &[u8]
) -> Result<(), Error>

Sends data to the USB in bulk.

It will divide the data in chunks.

fn usb_bulk_recv(
device_handle: &DeviceHandle,
endpoint: u8,
data: &mut [u8]
) -> Result<(), Error>

Receives data from the USB in bulk.

It will divide the data in chunks.